+++ src/appleseed/renderer/modeling/input/source.h
@@ -66,28 +66,28 @@ class Source
     // Evaluate the source at a given shading point.
     virtual void evaluate(
         TextureCache&               texture_cache,
-        const foundation::Vector2d& uv,
+        const foundation::Vector2f& uv,
         ScalarInput&                scalar) const;
     virtual void evaluate(
         TextureCache&               texture_cache,
-        const foundation::Vector2d& uv,
+        const foundation::Vector2f& uv,
         foundation::Color3f&        linear_rgb) const;
     virtual void evaluate(
         TextureCache&               texture_cache,
-        const foundation::Vector2d& uv,
+        const foundation::Vector2f& uv,
         Spectrum&                   spectrum) const;
     virtual void evaluate(
         TextureCache&               texture_cache,
-        const foundation::Vector2d& uv,
+        const foundation::Vector2f& uv,
         Alpha&                      alpha) const;
     virtual void evaluate(
         TextureCache&               texture_cache,
-        const foundation::Vector2d& uv,
+        const foundation::Vector2f& uv,
         foundation::Color3f&        linear_rgb,
         Alpha&                      alpha) const;
     virtual void evaluate(
         TextureCache&               texture_cache,
-        const foundation::Vector2d& uv,
+        const foundation::Vector2f& uv,
         Spectrum&                   spectrum,
         Alpha&                      alpha) const;
 
@@ -132,7 +132,7 @@ inline bool Source::is_uniform() const
 
 inline void Source::evaluate(
     TextureCache&                   texture_cache,
-    const foundation::Vector2d&     uv,
+    const foundation::Vector2f&     uv,
     ScalarInput&                    scalar) const
 {
     evaluate_uniform(scalar);
@@ -140,7 +140,7 @@ inline void Source::evaluate(
 
 inline void Source::evaluate(
     TextureCache&                   texture_cache,
-    const foundation::Vector2d&     uv,
+    const foundation::Vector2f&     uv,
     foundation::Color3f&            linear_rgb) const
 {
     evaluate_uniform(linear_rgb);
@@ -148,7 +148,7 @@ inline void Source::evaluate(
 
 inline void Source::evaluate(
     TextureCache&                   texture_cache,
-    const foundation::Vector2d&     uv,
+    const foundation::Vector2f&     uv,
     Spectrum&                       spectrum) const
 {
     evaluate_uniform(spectrum);
@@ -156,7 +156,7 @@ inline void Source::evaluate(
 
 inline void Source::evaluate(
     TextureCache&                   texture_cache,
-    const foundation::Vector2d&     uv,
+    const foundation::Vector2f&     uv,
     Alpha&                          alpha) const
 {
     evaluate_uniform(alpha);
@@ -164,7 +164,7 @@ inline void Source::evaluate(
 
 inline void Source::evaluate(
     TextureCache&                   texture_cache,
-    const foundation::Vector2d&     uv,
+    const foundation::Vector2f&     uv,
     foundation::Color3f&            linear_rgb,
     Alpha&                          alpha) const
 {
@@ -173,7 +173,7 @@ inline void Source::evaluate(
 
 inline void Source::evaluate(
     TextureCache&                   texture_cache,
-    const foundation::Vector2d&     uv,
+    const foundation::Vector2f&     uv,
     Spectrum&                       spectrum,
     Alpha&                          alpha) const
 {
