+++ src/appleseed/renderer/modeling/material/normalmappingmodifier.h
@@ -60,7 +60,7 @@ class NormalMappingModifier
 
     virtual foundation::Basis3d modify(
         TextureCache&                   texture_cache,
-        const foundation::Vector2d&     uv,
+        const foundation::Vector2f&     uv,
         const foundation::Basis3d&      basis) const APPLESEED_OVERRIDE;
 
   private:
